requirements.txt
asgiref==3.2.7
Babel==2.8.0
chardet==3.0.4
Django==3.0.6
django-cors-headers==3.2.1
django-csvimport==2.15
django-filter==2.2.0
django-localflavor==3.0.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
drf-extensions==0.6.0
drf-writable-nested==0.6.0
Markdown==3.2.2
phonenumbers==8.12.3
psycopg2==2.8.5
python-stdnum==1.13
pytz==2020.1
sqlparse==0.3.1
